+++ changelog.md
@@ -4,11 +4,10 @@
 
 ## Version 0.4.0
 
-> .osu file donwload !
+> .osu file download !
 
 - Added the BeatmapFile Model
 - Added OsuApi.get_beatmap_file()
-
 
 ## Version 0.3.1